Fix Cartesian view exception when tThWidth is None
A delta_tth value is required for PlanarDetector.make_powder_rings(),
even if the plane data tThWidth is None. Give it a default value
if the plane data tThWidth is None.

In my testing, it appears that different values of delta_tth here do
not affect the output, so perhaps an arbitrary delta_tth will work
fine.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>