Change "CalProgressDialog" to "ProgressDialog"
This progress dialog class is now being used for several different
types of asynchronous operations. It is no longer being used for
just calibration. Therefore, rename it to "ProgressDialog", and
make the class more general. The clients of the class will set
things such as the label text, the range, and potentially progress
updates.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>