Set load panel state defaults after clearing
The load panel was being intialized before the dummy images were loaded,
resulting in the load panel state being set and then cleared. This caused key
errors when the load panel settings were changed in any way. Instead, emit a
signal any time the load panel state is cleared so that the defaults are reset.

Signed-off-by: Brianna Major <brianna.major@kitware.com>