Don't switch to other items during renaming
It seems that the default QComboBox behavior with a QLineEdit is to
automatically change the index to another item, if there is another
item with the text that the user has typed in.

We don't want this behavior, but it is difficult to prevent. We are
preventing it here by intercepting the `QFocusEvent`, and changing
the index back to the original item that was chosen.

Now, if a user tries to rename a detector to be another item in the
list, when they press enter or lose focus, it reverts back to the
original detector name.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>