Increase the precision of floats in the tree view
Previously, when editing floating values in the tree view, the
floats would immediately be truncated to 2 decimal places when editing
began, and it would not allow you to specify more than 2 decimal places
of precision.

To fix this, we are using our ScientificDoubleSpinBox class for floats,
which allows for scientific notation. Additionally, the previous commit
increased the precision of the scientific spinbox itself. These changes
will hopefully provide enough precision for the tree view.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>