Draw material overlays in raw view
This utilizes PlanarDetector.make_powder_rings() to draw the material
overlays on the raw view.

It looks good for the most part, except that some detectors produce
ring_xys values (from PlanarDetector.make_powder_rings()) that wrap
around to the other side, so that a line is drawn across the detector
connecting the ends. I am not sure if this is something that should
be fixed in hexrdgui, or in hexrd.

Additionally, the values do not appear to go all the way to the image
borders for some reason.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>