Move materials table into a dialog
This moves the materials table into a dialog, which is not shown
by default. It can be shown by clicking the new "Show Materials Table"
button in the materials panel. The materials table can be hidden by
clicking on the "x" button in the top right corner.

All of the functionality should be the same. Rows are highlighted if
they are to not be excluded. The dialog will also automatically
update when settings are modified in the materials panel.

These changes were done to free up some space in the materials panel.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>