Create new HEDMInstrument class from config dict
A [new HEDMInstrument class](https://github.com/HEXRD/hexrd/blob/eb2f5ffd288287df9c331755b240d90a9bb7ea25/hexrd/instrument/instrument.py#L167) was added as a part of hexrd v0.8.
This HEDMInstrument class does not currently accept an instrument
config for initializing it, and this causes errors in hexrdgui
for the cartesian and polar viewers.

To adjust our code for this, add a new function for creating an
HEDMInstrument from the config, and use this when we need to create
an instrument.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>