Add -c anaconda to conda instructions
`-c anaconda` is added before `-c conda-forge` to give preference to
packages on anaconda if they are there, and then conda-forge if they
are not.

Hopefully, the packages on anaconda are more stable because they are
maintained by anaconda, rather than the community.

Additionally, numpy on anaconda uses mkl, and numpy on conda-forge
uses libopenblas. We have had problems with libopenblas on the mac,
so maybe mkl will be more stable.

Fixes: #298
Fixes: #299

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>