Return ranges from PowderLineOverlay as well
This modifies the PowderLineOverlay so that ranges and range indices
can be returned from the function as well. It does so by now returning
a dict that contains:

1. The rings
2. The ranges
3. The range indices

A large part of the `overlay()` function was moved into another function
so that it can be called multiple times for both the rings and the ranges.

A modification was also made so that the raw mode converts to pixel
coordinates, which is used in the raw images.

Finally, the overlay data is now also returned as a list of separate
overlays, rather than as a single nan-delimited list containing all
overlays. This better matches the behavior before, and it seems to
make more sense to have a separate matplotlib Line2D for each overlay.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>