Only draw rings once when they are shown
Rings were being drawn twice when they were made visible, due to two
different emits of HexrdConfig.ring_config_changed. This was especially
problematic for the raw view, where it was noticeably slow, since it
must draw overlays for each detector.

The performance seems to be improved after this change.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>