Auto-generate Cartesian params from config
When a new instrument config is loaded (either by
"File"->"Open"->"Configuration", or by resetting the instrument
configuration), automatically generate Cartesian parameters based
upon the config.

The parameters are currently calculated as follows:

1. Distance: absolute value of the average of detector z translation values
2. Pixel size: 5x the average of the detector pixel sizes

I ended up using 5x the average for the pixel size because that gave me a
reasonably interactive setup for my personal computer. But we can adjust it
if something else is better. Less than 5x, such as 3x, made for much slower
interaction for me.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>