Run the whole indexing and write results
The user can now run the whole indexing step, and the results get
written out to files in their current working directory.

In the find_orientations config:

1. The instrument gets automatically set from the hexrdgui instrument.
2. The materials dict gets automatically set from the hexrdgui config.
3. The active material gets automatically set from the hexrdgui config.
4. The active hkls get automatically set from the materials' planeData.

This eliminates some of the options that the user sets in the tree view.

Once the options are chosen in the tree view, and the user selects "Ok",
find_orientations is ran, the results are printed, and they are also
written out to a file.

Some improvements that will be made in the future:

1. Bring up a window that will display the log information as it comes.
2. Add some visualization.
3. Allow for user feedback at various steps.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>