Fix 'apply to all' behavior for transforms
In the case of a single frame image loaded through the load panel, toggling the
'Apply to all detectors' option would not un-check the box, but would toggle the
setting internally, resulting in unexpected behavior. Fix the behavior to allow
toggling the 'Apply to all' option and adjusting the results accordingly.

Signed-off-by: Brianna Major <brianna.major@kitware.com>