Generalize color map editor for multiple use
This generalizes the color map editor a little bit so that it can
be re-used in other parts of the program.

It now takes as an argument a general "image_object", and it defines
in the comments what functions/properties must be on the image_object.

This also removes where it is setting a value in the HexrdConfig, and
it uses nanpercentile instead of percentile (to take into account
nans), and it also allows the image data to be passed in as a dict,
tuple, list, or single array.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>