Add new overlay manager
This adds an overlay manager, where overlays can be added, edited,
and removed.

The overlay manager can be shown via the materials panel. Editing
an overlay makes the overlay editor appear. In the overlay editor,
there will be options specific to each type of overlay, and
there is the ability to edit the material, the type of overlay,
and the style (a button will open the overlay style editor).

This also saves the overlays in between sessions via QSettings.

This currently only works for powder. Soon, we will add the ability
to draw Laue overlays as well, and then Mono Rotation Series.

We can also re-think the design if there are too many dialog boxes.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>