Auto-compute polar resolutions from config
When an instrument config file is loaded, or the instrument config
is reset, the polar resolutions are automatically computed and set
based upon the config.

This is currently being done by scaling the max. However, we should
also consider median and min. We most likely need to do some
kind of scaling, so that we can come up with reasonable values for
most instrument configs.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>