Add Laue overlay widths to the polar view
This adds an option in the overlay editor to enable widths for the Laue
overlays, and it also adds options for tth_width and eta_width.

If this is enabled, these options get passed on to the Laue overlay
generator. For the polar view, these are used to construct range data,
which appear as rectangles around the Laue spots.

Currently, these options are ignored for the raw and cartesian views.
They will need to be added at a later time.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>