Use valueChanged, not editingFinished
editingFinished seems to be a little more unintuitive, in that, it
only appears to update when the widget loses focus. This means, clicking
the spin box arrows won't cause an update, until you press tab.
Additionally, an update still happens if the widget loses focus, even if
the value inside has not been modified.

valueChanged is more intuitive, because it only causes an update if the
value has been changed, and it also causes an update if the arrows are
clicked.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>