Remove distortion dict when 'None' is used
LLNL users will most often use a distortion of `None` when they use
the program. These options do not even need to be seen by them.
Thus, when the distortion is `None`, remove the distortion dictionary
entirely.

If the distortion gets turned back on to something, it now gets added
automatically back in.

This is also helpful for use in hexrd, which requires the distortion
dict to be missing if `None` is set.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>