Add calibration crystal editor with stretch matrix
This separates out the calibration crystal editor as a stand-alone
widget, because it will be re-used for the mono rotation series.

Additionally, the stretch matrix is used instead of an inverse
stretch matrix. This is automatically converted to inverse stretch
values behind the scenes, which then get passed to hexrd.

If the user inputs values for a stretch matrix that is a noninvertible
(singular) matrix, the matrix boxes appear as red, and the values
are not set. But once the matrix is invertible, the values will
appear as white again.

Additionally, values across the diagonal are kept identical in the
stretch matrix automatically.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>