Floor number of pixels for polar plot
Previously, there would sometimes be white space on the right
side of the polar plot due to extra pixels being generated. Floor
the number of pixels so this doesn't happen. From our examples, it
looks like this doesn't remove any noticeable part of the plot.

This also gets the polar mask function to use the same logic as
the polar view, which is required to ensure that the numbers of
pixels are the same for both.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>