Allow manual association
If the user has not selected all of the images they need and the rest cannot be
found by matching files in the same or sibling directories, assume that the user
will manually associate each file with its respective detector. Provide them with
the option to select each detector in the file confirmation dialog.

Signed-off-by: Brianna Major <brianna.major@kitware.com>