Add default options for overlay types
Previously, if there were no options provided, the overlay generator
would set defaults. These defaults would be different from the GUI
defaults, which led to a mismatch.

To avoid this, create default options for the different overlay types,
and always set those default options when the overlay is created or
the type is changed. This ensures that the overlay generator and the
GUI both use these defaults, to avoid a mismatch.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>