Simplify zoom canvas
This adds simplifications to many parts of the zoom canvas, which
results in a nice speedup.

One of the simplifications is that we are skipping drawing a polygon
for the rectangle, and just using numpy indexing directly.

Another simplification is that we draw the whole image the first
time, and then moving the cursor just changes the zoom, rather than
drawing a new image each time.

The integration plot is now also autoscaled.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>