Move material tth width editor into powder overlay
Rather than having the material tth width editor be on the
materials panel, move it into the powder overlay editor, because
it currently only impacts the powder overlays.

This tth width is still set on the material, so two powder overlays
with the same material will have the same tth width, but that is
most likely not an issue.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>