Clean up polar view autoscaling
Previously, the azimuthal integral axis would be autoscaled once, and
then the autoscaling would be turned off. This is not necessary, and
it interferes with manual autoscaling in the future.

Remove this code, and when the azimuthal integral gets updated,
manually autoscale the y-axis only. This fixes automatic rescaling
of the x-axis when zoomed in to the polar view.

This commit also removes some no-longer-needed checks for the x-axis
scaling, and it re-arranges a few parts for autoscaling of the
main axis as well. The code seems to be cleaner and more logical, and
it seems to also function as intended.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>