Prevent setData(...) being called when button is clicked
When the QPushButton triggers a FocusOut event, such when
we programmatically disable it is causes setData(...) to
be called on the model with the value of the button ( as in
if its enabled or disabled ). I am not sure of the reason
for this, I couldn't find anything in the documentation.
To workaround this we install an event filter, to prevent
this event from propagating when we disable or enable the
button.