Add a structure tab to the materials panel
The structure tab modifies sites and atomic properties for the
material.

Only a single U value for each atom type is currently allowed. However,
in the near future, we may need to allow for a symmetric tensor for
the U values for each atom type.

Currently, to update the `hexrd` objects, it updates the attributes directly
on the `Material` class, and then re-creates the `unitcell` from scratch, since
that seems to be easier than updating all the properties inside the `unitcell`.
It then calls a function to update the structure factor on the `PlaneData` object.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>