Add WPPF generation and plot to the polar view
This adds a new menu item, "Run"->"WPPF", which makes an options
dialog appear, and then generates WPPF data using the active
material, current beam energy, selected options, and the current
azimuthal integral plot.

Currently, only the "LeBail" WPPF method works. "Rietveld" is an option
but does not currently work.

Additional options appear in the dialog if the user selects a background
method that uses additional parameters. The user can either choose to
use an experiment file for the fitting data, or the current azimuthal
integral plot data will be used.

When the azimuthal integral axis in the polar view is drawn, the image
canvas automatically checks for the presence of WPPF data. If any was
generated, it will automatically plot it. There is currently no way
to clear the WPPF data (you can only replace it with a new one). So
once it has been generated, it will stick around. We can probably make
certain operations clear it at some point.

Currently, there can only be one set of WPPF data at a time. Also, the
options in the dialog are saved and re-loaded during the current run,
but are not saved in between runs. It is an easy fix if we wish to save
the options in between runs.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>