Allow 'inf' and '-inf' in ScientificDoubleSpinBox
These correspond to python's `math.inf` and `-math.inf`, respectively.
They can be alternatively obtained via `float('inf')` and `float('-inf')`.

If there is a max and min set on the spinbox, the value will be clipped
to the max and min. If not, however, `inf` and `-inf` are allowed and valid.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>