Average x for azimuthal integral plot overlays
This fixes an issue where there are overlays for every x value,
which is problematic if the powder offset is not at the origin,
because the line overlays are curved.

There may be alternatives to averaging that may be preferred. But
we'll do that if we find a better alternative. For now, this fixes
the extremely slow GUI when the line overlays are curved.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>