Save Laue spot data in the polar view as lists
Instead of numpy arrays.

This is done because numpy arrays do some fancy optimizations,
which make it so that things like
`[id(x) for x in spots]` does not return predictable results.
Sometimes, for instance, a new view object is returned,
which will have a different id than the object with the data.

However, if we use lists instead, the `id()` of the object
is predictable. We are going to use the `id()` of the spots
to keep track of highlighting, at least for now.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>