Add ability to select materials for WPPF
This adds a button the WPPF options dialog that will pop up a dialog
where the user can select which materials to use for WPPF.

Only materials with powder overlays can be used. Materials that are
both visible and have powder overlays are selected by default. Materials
that have powder overlays but are not visible can be selected in the
dialog.

These materials then get passed on to WPPF. Previously, only the currently
selected material was passed to WPPF. Multiple materials can now be passed.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>