Add ability to specify crystal refinements
This adds a "refinement" tab to the calibration crystal editor to
allow the user to specify which crystal parameters can be refined.

It is currently being used only in the Laue overlay editor.

The refinement statuses are saved in the overlays themselves. They
are currently not being used for anything, but they will be used when
the spot picked calibration will be performed.

All parameters default to being refinable.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>