Add general calibration runner
This generalizes the line picked calibration into a general calibration
runner, with several helpful features.

First of all, the calibration runner loops through all of the visible
overlays, and has the user pick points for each one. While the user
is picking points for an overlay, all of the other overlays are hidden,
and the line/spot of interest is highlighted.

For powder overlays, the user can move on to the next line by
right-clicking. For laue overlays, right-clicking is disabled and
left click automatically moves on to the next spot. If a spot in the
laue overlays needs to be skipped, the user may do so by middle clicking,
which inserts NaN's instead of regular values.

The user may click on the "Ok" button on the dialog to move on to the next
overlay. Alternatively, if there are no more points available for picking,
the dialog will automatically close itself.

Currently, nothing is being done with the picked data. But it will be fed
into a calibration function soon.

For powder overlays, the detector association may not be correct if the
user did not click on the right detector. To make things simpler, we should
just automatically determine which detector each point is on and associate
them in that way.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>