Exit early in powder overlays if there is no data
Check and exit early if there is no data when generating the powder
overlays.

Sometimes, there is no overlay data just because there are exclusions
on all of the hkls. Why this happens sometimes in currently unknown.
But this should alleviate that issue.

Sometimes, there is no overlay data because the max two theta in the
materials panel and the beam energy are both too small. This results
in no hkls appearing in the materials table.

This early exit prevents exceptions from occurring, and instead, no
powder overlays are drawn, which is more expected behavior.

Fixes: #500

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>