Save image config to QSettings
This will result in the image config being preserved in between
hexrdgui runs. This is especially useful for cases where the
polar max two theta is severely truncated in between runs compared
to what it was before, which also results in a small max two theta
in the materials panel, which results in no active hkl values.

This also moves updating the snip1d enable states to a function
so we can also call it when the GUI gets updated.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>