Make total occupancy computed instead of editable
The total occupancy is now disabled for editing, and it is now
computed automatically from the sum of the occupancies, rather than
being editable by the user.

Now, the only validation error that occurs is if the total occupancy
is greater than 1, which is not allowed.

In addition, U values are kept when possible when the atom type list
is edited. But the individual occupancies are always reset when the
atom type list is edited.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>