Always update polar view when beam vector changes
The polar view was only being updated for changes to the beam vector
in the slider widget, but not for changes to the beam vector in the
form view or tree view.

It now updates for all three views. There is also a timer attached to
it so that if the slider view is being used, it doesn't trigger too
many updates.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>