Update overlays in raw view mode on detector edit
When the detector transforms get modified, update the overlays
in the raw view.

This is a little slow to use in the slider view, because it is
currently re-generating overlays for all of the raw images (not just
the one that got changed). But maybe we can fix that in the future...

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>