Run calibration and update hexrdgui with output
This seems to work for the most part, except that the material
lattice parameters do not seem to be getting refined (it looks like
the lattice params output from the calibrator is always the same,
even though it is flagged as refinable).

Note that the user currently has to step through all overlays. They
cannot click the "Ok" button to exit early, or there will be errors
from the calibrator about list sizes not matching. This is hopefully
something we can fix in the future.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>