Don't clear axes on color by change
Clearing the axes was being performed to remove the scatter plot artist.
However, it was problematic because it was also removing the axes labels.

Instead of clearing the axes, remove the scatter plot artist manually.
I unfortunately could not find any way to update the colors on the scatter
plot object, nor the mappable on the colorbar. So both need to be re-created
from scratch each time the color changes.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>