Add a ranges editor
This allows editing x, y, and z ranges, and it also allows
the user to reset the ranges.

When the min is greater than the max, the behavior in matplotlib
is a little odd (the interaction gets reversed, for one thing, and
the side that zero is one gets moved). To prevent this, we are
now allowing the min to be greater than the max.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>