Add home, back, and forward buttons
In theory, they should work with the latest matplotlib. In reality,
they are buggy still. However, we'll add them in anyways.

In order for our ranges editor to stay in sync with the underlying
matplotlib ranges, some callbacks were added to update the ranges
editor after home, back, and forward are performed.

This also updates the navigation stack when the matplotlib ranges
are modified, so that the home/back/forward buttons can be used
for modifying ranges.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>