Add a generic matrix editor widget
This matrix editor widget is passed a 2-D numpy array, and it
constructs a matrix of spinboxes based on the shape of the numpy array.
Modifying the spinboxes causes the numpy array to be modified directly.
A signal is emitted that indicates when the array has been modified.

There are additionally ways to specify which elements of the matrix are
editable, and you can also supply a constraint function to ensure that
constraints are satisfied when elements in the matrix are edited.
The constraints function should force a disabled element to be equal
to an enabled element.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>