Reset disabled matrix values to zero after update
Some disabled matrix values are disabled because they are constrained
to match another value. Some are disabled because they are zero.

When the enable states change, reset all disabled values to zero, then
apply constraints, and update the GUI. This prevents disabled values
that should be zero from having a non-zero value.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>