Edit indexing and fit grains configs via tree view
This adds editing for indexing and fit grains configs via a tree
view. These are located under the "View" menu item.

Right now, you can edit all of the values in the right column,
which lets you edit most of the config options. However, you cannot
edit the keys. For the seed search method, you need to be able to
change the key in order to change the method, and this in turn
changes children keys. This is not supported currently, and you
can only currently edit values.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>