Update spots with chosen method in ome maps viewer
Previously, it always used logic for the "label" method when creating
the spots, and it wasn't even fully accurate for the "label" method (it
was missing some of the code logic).

Instead, we should call the "label_spots" function in hexrd so the spots
are generated the exact same way that hexrd does, to give the user an
accurate preview.

Additionally, it now generates spots based upon the method that has
been selected (label, blob dog, or blob log), and the options for each of
those methods. The spots are automatically updated with any change to
the method options or to the method itself.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>