Use masked indices for active_hkls in indexing
The "active_hkls" are supposed to use masked indices (indices for
the hkl data list where the exclusions and tThMax have already been
applied). Before, they were trying to use unmasked indices, which
refer to the hkl data list where exclusions and tThMax have not been
applied.

It appears this convention is done nearly everywhere in hexrd, so
we should do it too.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>