Update dialog
Add the option to select files manually or to enter a search pattern to grep with
in the set directory, as well as the option to specifiy if the search should
apply to all directories or just the current one. The dialog was becoming quite
long and crowded, so this also rearranges things to sit together a bit more
logically.

Signed-off-by: Brianna Major <brianna.major@kitware.com>