Support multiple load options for files
Files can now be selected manually for each detector or they can be searched for
using a regex. If searching using a file pattern the user can either provide a
search for each detector or apply one search to all detectors. This also checks
that directories have been set for and that the same number of files have been
selected/found for each detector.

Signed-off-by: Brianna Major <brianna.major@kitware.com>