Update structure factor on size mismatch
If we save a materials file that has some exclusions, then when
that materials file is loaded, the structure factor that the material
generates will exclude the hkls in the exclusion list. This results
in a mismatch for the GUI when it tries to load all hkls without
exclusions.

Therefore, check if there is a size mismatch, and update the structure
factor if that is the case.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>