Avoid trying to format MaskedArray instances
Matplotlib is creating a MaskedArray from the data, where nans and infs
are masked. We tried to format it with a float formatting string, and
this was producing a warning on Linux and an error on Windows.

Check to see if the intensity is an instance of np.ma.MaskedArray
(as a side note: np.ma.core.MaskedConstant inherits from np.ma.MaskedArray,
so this check also works for np.ma.core.MaskedConstant). If it is,
then show another label and don't try to format it.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>