Perform filtering and cleaning at the right times
Hexrd performs the cleaning in generate_orientation_fibers right
now, so we should not perform the cleaning. Instead, create a deep
copy in the viewer and perform the cleaning for the visualization,
and let hexrd perform the cleaning on its own later.

Hexrd also normally filters the maps immediately after they are
loaded, which we are not doing. So we must filter the maps after
the user has selected the options.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>