Experimental: remove make_new_pdata
It seems that often (perhaps always), creating a new PlaneData object
from scratch is not necessary, and is a performance hit.

Try removing all cases of re-creating a PlaneData object from scratch,
and instead, we should ensure that when the relevant hexrd Material
attributes get modified, that the underlying PlaneData object will
get properly updated as well (including anything that needs to be
re-computed).

This currently seems to give a nice performance boost for modifying
the energy via the slider widget, and seeing the overlays move.

However, we're going to need to change some of the properties in
hexrd to update everything that is needed, and we should ensure
that everything in the GUI is updating as expected.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>