Materials panel update cleanup
This cleans up the materials panel updating system to make it
easier to keep track of what needs to be updated and when.

Now, when the active material gets changed, a signal fires
automatically from HexrdConfig() and the appropriate update functions
get called in the materials panel.

Additionally, when the active material gets modified, the appropriate
update functions also get called.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>