Apply 2D panel buffer as a mask to the polar view
This checks if the panel buffer is a 2D numpy array, and if it is,
it will apply it as a mask to the polar view.

This seems to correctly generate NaNs in the output. However, the
image doesn't look so good. Maybe the mask is messing up the math
somewhere...

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>