Add automatic powder calibration
This performs automatic calibration for a single powder overlay.

It first automatically picks points, and then provides visual feedback
in the canvases so that the users can see the points that were chosen.

The user is then able to accept or reject the automatically picked points.

This is taking over the "Powder Calibration" menu item. It is similar, but
it also works for merged ranges, and it provides visual feedback.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>