Use only required params for default refinements
Before, it was adding all of the parameters to the default refinements.

Now, it should be making better defaults based upon which lattice
parameters are required.

Now, the default overlay refinements function accepts a whole overlay
dict, and it extracts from it the type and the material name in order
to come up with better default refinements.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>