Save/load .hexrd instrument config files
These are HDF5 files that hexrd is now capable of saving/loading.

This is now the default instrument config file type, but yaml files
can still be loaded as well.

Now also, the way the files are saved/loaded internally has been changed.
Before, a dict was just written out to a yaml file. Now, an HEDMInstrument
object is created and used to write out either the yaml/hdf5 file.

Also, save/load of .hexrd directories is no longer supported, since it is
not needed anymore.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>