Check for np.ndarray when making status lists
It looks like in some cases, the instrument config value lists can be
np.ndarray now. We noticed this specifically for distortion parameters.

Add a check for np.ndarray as well when creating statuses, so that a
list of statuses will be properly created when it receives an np.ndarray.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>