Add basic refinements editor
This adds a basic refinements editor under the "Edit"->"Refinements"
menu item. It allows the user to edit both the refinements and values
of refinable parameters that are used for calibration.

It additionally comes with a combo box of actions to automatically set
refinements. Currently, the only action is "Mirror first detector", which
causes the refinement settings of all of the detectors to match what is
currently set for the first detector.

It is hopefully made in a way so it will be easily extensible to include
additional things such as lower bounds and upper bounds, and other
types of constraints.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>