Add a context manager utility for blocking signals
This is to make our use of QSignalBlocker more pythonic. Instead
of relying on the QSignalBlocker object to be destroyed and having
a "# noqa: F841" to fix the flake8 issues, we can now use a with
block instead, like:

```python
with block_signals(obj1, obj2, ...):
    # Do something...
    ...
```

This is the more pythonic way to acquire and release resources.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>