Use an interpolation of 'none' for raw view
It defaults to use anti-aliasing, which is very slow. We have been using
an interpolation of 'none' for both cartesian and polar views. For now,
let's try to use an interpolation of 'none' for the raw view as well,
as it seems to render the images much faster.

This will help a lot with the B&C editor.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>