Always use draw_idle() in the image canvas
draw_idle() gets the GUI to update the next time control is returned
to the event loop, and multiple calls to draw_idle() will result in
just one call to draw().

This is a nice and easy way to improve the efficiency of our drawing,
and we should definitely look into using it in other places as well.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>