Add a "Reflections Table" button to powder editor
This will bring up the reflections table for the material that is
currently being edited in the powder editor.

This commit also adds some new mechanics so that all reflections tables
matching the active material will update when the materials panel is
edited. This will hopefully keep them all up-to-date.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>